updating the api and hook a little bit
